backend/ia-admin/requirements.txt
# 🐍 AGROISYNC IA - Dependências Python

# Framework Web
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Segurança e Autenticação
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP e Requests
httpx==0.26.0
requests==2.31.0

# Validação de Dados
pydantic==2.5.3
email-validator==2.1.0

# Utilitários
python-multipart==0.0.6

